move interfaces to interfaces folder
