+++ src/SFA.DAS.FindEmploymentSchemes.UnitTests/Web/Controllers/PagesControllerTests.cs
@@ -7,7 +7,7 @@ using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content.Interfaces;
 using Xunit;
 using SFA.DAS.FindEmploymentSchemes.Web.Controllers;
 using SFA.DAS.FindEmploymentSchemes.Contentful.Services;
-using SFA.DAS.FindEmploymentSchemes.Web.Services;
+using SFA.DAS.FindEmploymentSchemes.Web.Services.Interfaces;
 
 namespace SFA.DAS.FindEmploymentSchemes.UnitTests.Web.Controllers
 {
