+++ src/SFA.DAS.FindEmploymentSchemes.Web/Services/Interfaces/IFilterService.cs
@@ -1,7 +1,7 @@
 ﻿using SFA.DAS.FindEmploymentSchemes.Web.Models;
 using SFA.DAS.FindEmploymentSchemes.Web.ViewModels;
 
-namespace SFA.DAS.FindEmploymentSchemes.Web.Services
+namespace SFA.DAS.FindEmploymentSchemes.Web.Services.Interfaces
 {
     public interface IFilterService
     {
