+++ src/SFA.DAS.FindEmploymentSchemes.Web/Services/Interfaces/IPageService.cs
@@ -1,7 +1,7 @@
 ﻿using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content;
 using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content.Interfaces;
 
-namespace SFA.DAS.FindEmploymentSchemes.Web.Services
+namespace SFA.DAS.FindEmploymentSchemes.Web.Services.Interfaces
 {
     public interface IPageService
     {
