+++ src/SFA.DAS.FindEmploymentSchemes.Web/Services/Interfaces/ISchemesModelService.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.FindEmploymentSchemes.Web.Models;
 
-namespace SFA.DAS.FindEmploymentSchemes.Web.Services
+namespace SFA.DAS.FindEmploymentSchemes.Web.Services.Interfaces
 {
     public interface ISchemesModelService
     {
