+++ src/SFA.DAS.FindEmploymentSchemes.Web/Services/PageService.cs
@@ -3,6 +3,7 @@ using System;
 using System.Linq;
 using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content;
 using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content.Interfaces;
+using SFA.DAS.FindEmploymentSchemes.Web.Services.Interfaces;
 
 namespace SFA.DAS.FindEmploymentSchemes.Web.Services
 {
